move ManagedDialogProtocol to api
