+++ aiogram_dialog/api/protocols/__init__.py
@@ -1,5 +1,7 @@
 __all__ = [
+    "ManagedDialogProtocol",
     "MediaIdStorageProtocol",
 ]
 
+from .managed import ManagedDialogProtocol
 from .media import MediaIdStorageProtocol
